refactor(sdk): skill-facing exceptions live in innate.exceptions
Per review: ArmFailed/ArmUnhealthy move out of manipulation.py into
brain_client.robot.exceptions, re-exported with SkillFailed/SkillCancelled
from innate.exceptions (and at the innate top level). Old import paths
keep working.
