+++ ros2_ws/src/brain/brain_client/brain_client/robot/exceptions.py
@@ -1,0 +1,11 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (c) 2026 Innate Inc
+"""Robot interface exceptions, ROS-free so ``innate.exceptions`` can re-export them."""
+
+
+class ArmUnhealthy(RuntimeError):
+    """Servo brownout/trip/refusal — abort, don't continue limp."""
+
+
+class ArmFailed(RuntimeError):
+    """Arm command rejected, unreachable, or did not complete."""
