+++ ros2_ws/src/brain/brain_client/innate/__init__.py
@@ -49,10 +49,13 @@ interface helpers) raises SkillCancelled the moment a Stop lands, the base
 is braked automatically, and the run reports CANCELLED. ``try/finally`` in
 execute() is your cleanup hook; ``self.on_cancel(...)`` exists only to
 forward a cancel to an external action goal.
+
+:mod:`innate.exceptions` groups every exception a skill raises or catches.
 """
 
 from typing import TYPE_CHECKING
 
+from brain_client.robot.exceptions import ArmFailed, ArmUnhealthy
 from brain_client.skills.types import (
     PhysicalSkill,
     Skill,
@@ -113,18 +116,14 @@ __all__ = [
 # Type checkers can't follow __getattr__, so they read the imports below.
 if TYPE_CHECKING:
     from brain_client.robot.head import Head
-    from brain_client.robot.manipulation import ArmFailed, ArmUnhealthy, Manipulation, Waypoint
+    from brain_client.robot.manipulation import Manipulation, Waypoint
     from brain_client.robot.mobility import Mobility
 
 _LAZY_INTERFACES = {
     "Mobility": ("brain_client.robot.mobility", "Mobility"),
     "Manipulation": ("brain_client.robot.manipulation", "Manipulation"),
     "Head": ("brain_client.robot.head", "Head"),
-    # Manipulation's command vocabulary: the trajectory waypoint and the
-    # exceptions its motion methods raise.
     "Waypoint": ("brain_client.robot.manipulation", "Waypoint"),
-    "ArmFailed": ("brain_client.robot.manipulation", "ArmFailed"),
-    "ArmUnhealthy": ("brain_client.robot.manipulation", "ArmUnhealthy"),
 }
